displayed our first chart using Chartjs
