--- client/vite-project/src/components/Graph.js
@@ -1,35 +1,0 @@
-import React from "react";
-import { Doughnut } from 'react-chartjs-2';
-import { Chart } from 'chart.js';
-
-export default function Graph(){
-   return(
-      <div className="flex justify-content max-w-xs mx-auto">
-       <div className="item">
-          <div className="chart relative">
-            Chart      
-        </div> 
-        <div className="flex flex-col py-10 gap-4">
-         {/* Labels */}
-        </div>
-       </div>
-        </div>
-   ) 
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
